feat: user-region entity 생성
