+++ src/main/java/com/example/sportspie/bounded_context/auth/entity/User.java
@@ -1,4 +1,4 @@
-package com.example.sportspie.auth.entity;
+package com.example.sportspie.bounded_context.auth.entity;
 
 import java.util.Collection;
