+++ src/main/java/com/example/sportspie/bounded_context/auth/entity/UserRegion.java
@@ -1,6 +1,5 @@
-package com.example.sportspie.auth.entity;
+package com.example.sportspie.bounded_context.auth.entity;
 
-import com.example.sportspie.auth.type.Position;
 import com.example.sportspie.base.entity.BaseTimeEntity;
 
 import jakarta.persistence.Column;
@@ -8,7 +7,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -17,17 +15,11 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserPosition extends BaseTimeEntity {
+public class UserRegion extends BaseTimeEntity {
 	@JoinColumn(name = "user_id", nullable = false)
 	@ManyToOne
 	private User user;
 
-	@Column(columnDefinition = "TINYINT", nullable = false)
-	private Position position;
-
-	@Builder
-	public UserPosition(User user, Position position) {
-		this.user = user;
-		this.position = position;
-	}
+	@Column(nullable = false)
+	private String region;
 }
